Use own store for shortcuts (allows shortcuts for all functionalities)
